requirements.txt
# requirements.txt
Flask==3.0.3
requests==2.32.3
paho-mqtt==1.6.1
python-dotenv==1.0.1